Create AlarmClock.py
This code prompts the user to enter the desired alarm time and then generates a random math problem. When the alarm time is reached, the program asks the user to solve the math problem to switch off the alarm. If the solution is correct, the alarm is switched off; otherwise, a new math problem is generated until the correct solution is provided.